pyproject.toml
[tool.poetry]
name = "rankedle"
version = "0.1.0"
description = ""
authors = ["Luis Escolano <luis@escolano.es>"]
readme = "README.md"
packages = [{include = "rankedle", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.2"
asyncio = "^3.4.3"
pydub = "^0.25.1"
pyyaml = "^6.0.2"
yt-dlp = "^2025.1.15"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
